requirement.txt
ipykernel
python-dotenv
google
numpy
faiss-cpu
streamlit
requests